feat: :sparkles: Dragging item to another sheet while pressing Ctrl will cause the item to be copied over

+++ module/sheets/actor-sheet.mjs
@@ -2008,7 +2008,7 @@ export class SdmActorSheet extends api.HandlebarsApplicationMixin(sheets.ActorSh
       return [];
     }
 
-    if (itemData[0].parent?.isOwner) {
+    if (itemData[0].parent?.isOwner && !event.ctrlKey) {
       const item = fromUuidSync(itemData[0].uuid);
       await item.delete();
     }
